requirements.txt
certifi==2018.1.18
chardet==3.0.4
colorama==0.3.9
Django==2.0.1
django-filters==0.2.1
djangorestframework==3.7.7
httpie==0.9.9
idna==2.6
Pillow==5.0.0
Pygments==2.2.0
PyJWT==1.5.3
PyMySQL==0.8.0
pytz==2017.3
requests==2.18.4
urllib3==1.22
